replace sizeof(size_t) with sizeof(uint32_t)

+++ Source/LemmatizerLib/MorphAutomat.cpp
@@ -21,10 +21,10 @@ static int  InitAlphabet(MorphLanguageEnum Language, int* pCode2Alphabet, int *p
 						&&	(AdditionalEnglishChars.find(i) != std::string::npos)
 					)
 				||	(		(Language == morphGerman) 
-						&&	(AdditionalGermanChars.find(i) != std::string::npos)
+						&&	(AdditionalGermanChars.find((BYTE)i) != std::string::npos)
 					)
 				||	(		(Language == morphURL) 
-					  &&	is_alpha(i, morphURL)
+					  &&	is_alpha((BYTE)i, morphURL)
 					 )
 			)
 		{
